fix some dependencies and includes; misc other stuff
