+++ app/io.hpp
@@ -39,10 +39,11 @@
 #pragma once
 #include "jet_defs.h"
 #include "jet_config.h"
-#include "contract.hpp"
 #include <sstream>
 #include <string>
 #include <iostream>
+#include <fstream>
+#include <iomanip>
 
 namespace jet_partitioner {
 
@@ -53,7 +54,7 @@ bool load_config(jet_partitioner::config_t& c, const char* config_f) {
         std::cerr << "FATAL ERROR: Could not open config file " << config_f << std::endl;
         return false;
     }
-    std::string lines[4];
+    std::string lines[5];
     int reads = 0;
     // you might think that reading in four lines from a simple config file could be done like:
     // f >> c.coarsening_alg; 
@@ -63,11 +64,11 @@ bool load_config(jet_partitioner::config_t& c, const char* config_f) {
     // but that doesn't work if there are exactly 3 lines instead of 4 in the config file
     // because if the last line is a float like 3.14, then c.num_iter will contain the 3
     // and the c.max_imb_ratio will contain the .14
-    for(int i = 0; i < 4; i++){
+    for(int i = 0; i < 5; i++){
         if(f >> lines[i]) reads++;
     }
     f.close();
-    if(reads != 4){
+    if(reads < 4){
         std::cerr << "FATAL ERROR: Config file has less than 4 lines" << std::endl;
         return false;
     }
@@ -75,6 +76,7 @@ bool load_config(jet_partitioner::config_t& c, const char* config_f) {
     c.num_parts = std::stoi(lines[1]);
     c.num_iter = std::stoi(lines[2]);
     c.max_imb_ratio = std::stod(lines[3]);
+    if(reads == 5) c.ultra_settings = std::stoi(lines[4]);
     return true;
 }
 
@@ -239,72 +241,6 @@ part_vt load_part(ordinal_t n, const char *fname){
     ifp.close();
     Kokkos::deep_copy(part_d, part);
     return part_d;
-}
-
-//loads a sequence of coarse graphs and mappings between each graph from binary file
-//used to control for coarsening when experimenting with refinement
-std::list<typename jet_partitioner::contracter<matrix_t>::coarse_level_triple> load_coarse(){
-    using coarse_level_triple = typename jet_partitioner::contracter<matrix_t>::coarse_level_triple;
-    std::list<coarse_level_triple> levels;
-    std::list<coarse_level_triple> error_levels;
-    FILE* cgfp = fopen("coarse_graphs.out", "r");
-    int size = 0;
-    if(fread(&size, sizeof(int), 1, cgfp) != 1) return error_levels;
-    std::cout << "Importing " << size << " coarsened graphs" << std::endl;
-    ordinal_t prev_n = 0;
-    for(int i = 0; i < size; i++){
-        coarse_level_triple level;
-        level.level = i + 1;
-        ordinal_t N = 0;
-        if(fread(&N, sizeof(ordinal_t), 1, cgfp) != 1)  return error_levels;
-        edge_offset_t M = 0;
-        if(fread(&M, sizeof(edge_offset_t), 1, cgfp) != 1) return error_levels;
-        edge_vt rows("rows", N + 1);
-        auto rows_m = Kokkos::create_mirror_view(rows);
-        if(fread(rows_m.data(), sizeof(edge_offset_t), N + 1, cgfp) != static_cast<size_t>(N + 1)) return error_levels;
-        Kokkos::deep_copy(rows, rows_m);
-        vtx_vt entries("entries", M);
-        auto entries_m = Kokkos::create_mirror_view(entries);
-        if(fread(entries_m.data(), sizeof(ordinal_t), M, cgfp) != static_cast<size_t>(M)) return error_levels;
-        Kokkos::deep_copy(entries, entries_m);
-        wgt_vt values("values", M);
-        auto values_m = Kokkos::create_mirror_view(values);
-        if(fread(values_m.data(), sizeof(value_t), M, cgfp) != static_cast<size_t>(M))  return error_levels;
-        Kokkos::deep_copy(values, values_m);
-        graph_t graph(entries, rows);
-        matrix_t g("g", N, values, graph);
-        level.mtx = g;
-        wgt_vt vtx_wgts("vtx wgts", N);
-        auto vtx_wgts_m = Kokkos::create_mirror_view(vtx_wgts);
-        if(fread(vtx_wgts_m.data(), sizeof(value_t), N, cgfp) != static_cast<size_t>(N)) return error_levels;
-        Kokkos::deep_copy(vtx_wgts, vtx_wgts_m);
-        level.vtx_w = vtx_wgts;
-        if(level.level > 1){
-            vtx_vt i_entries("entries", prev_n);
-            auto i_entries_m = Kokkos::create_mirror_view(i_entries);
-            if(fread(i_entries_m.data(), sizeof(ordinal_t), prev_n, cgfp) != static_cast<size_t>(prev_n)) return error_levels;
-            Kokkos::deep_copy(i_entries, i_entries_m);
-            typename jet_partitioner::contracter<matrix_t>::coarse_map i_g;
-            i_g.coarse_vtx = N;
-            i_g.map = i_entries;
-            level.interp_mtx = i_g;
-        }
-        prev_n = N;
-        levels.push_back(level);
-    }
-    fclose(cgfp);
-    return levels;
-}
-
-//reads part from binary file
-part_vt load_coarse_part(ordinal_t n){
-    FILE* cgfp = fopen("coarse_part.out", "r");
-    part_vt part("part", n);
-    auto part_m = Kokkos::create_mirror_view(part);
-    if(fread(part_m.data(), sizeof(part_t), n, cgfp) != static_cast<size_t>(n)) return part;
-    Kokkos::deep_copy(part, part_m);
-    fclose(cgfp);
-    return part;
 }
 
 };
